requirements.txt
Django>=5.0,<6
djangorestframework>=3.15.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
openai>=1.40.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0
scikit-learn>=1.4.0
gunicorn>=22.0.0
whitenoise>=6.7.0
